Added schedules to shuttleFactories and auto-calculate stop times in shuttle creation
