+++ components/modals/planificator/shuttle.vue
@@ -30,19 +30,16 @@
 
       <ec-field
         class="column"
-        label="Dates"
+        label="Jour et heure de départ"
         field-id="start"
       >
         <date-time
+          v-model="start"
           lang="fr"
           input-class="input"
           type="datetime"
-          range
-          :value="range"
           :minute-step="5"
           format="YYYY-MM-DD HH:mm"
-          range-separator="->"
-          @input="updateDates"
         >
           <template slot="calendar-icon">
             <fa-icon icon="calendar" />
@@ -86,14 +83,15 @@
     <div class="select-shuttle-factory">
       <select-shuttle-factory
         id="shuttleFactory"
-        v-model="shuttle.shuttleFactory"
+        :value="shuttle.shuttleFactory"
         :disabled="shuttle.shuttleFactory.stops.length > 0"
+        @input="selectShuttleFactory"
       />
     </div>
 
     <div class="columns">
       <stops-manager
-        v-if="shuttle.shuttleFactory.stops.length > 0"
+        v-if="shuttle.shuttleFactory.stops && shuttle.shuttleFactory.stops.length > 0"
         :parent-stops="shuttle.shuttleFactory.stops"
         :parent-passengers="shuttle.passengers"
         :capacity="capacity"
@@ -124,9 +122,7 @@
 </template>
 
 <script>
-import { DateTime } from 'luxon';
 import cloneDeep from 'lodash.clonedeep';
-
 import vueModal from '~/components/modals/default.vue';
 import selectShuttleFactory from '~/components/form/selects/shuttle-factories.vue';
 import searchAvailableDriver from '~/components/form/search-available-driver.vue';
@@ -134,6 +130,7 @@ import searchAvailableCar from '~/components/form/search-available-car.vue';
 import stopsManager from '~/components/modals/planificator/shuttle/stops-manager.vue';
 
 import stopsHandler from '~/components/modals/planificator/shuttle/mixins/stops-handler';
+import passengersHandler from '~/components/modals/planificator/shuttle/mixins/passengers-handler';
 import toggleModal from '~/components/modals/planificator/shuttle/mixins/toggle-modal';
 import toggleLoading from '~/helpers/mixins/toggle-loading';
 import errorsManagement from '~/helpers/mixins/errors-management';
@@ -153,6 +150,7 @@ export default {
     toggleModal(),
     toggleLoading(),
     errorsManagement(),
+    passengersHandler(),
   ],
   props: {
     active: {
@@ -182,18 +180,6 @@ export default {
     ApiShuttle() {
       return this.$api.query('shuttles').setMask(mask);
     },
-    capacity() {
-      if (this.shuttle.car && this.shuttle.car.model && this.shuttle.car.model.capacity) {
-        return this.shuttle.car.model.capacity;
-      }
-      return 0;
-    },
-    range() {
-      return [
-        this.shuttle.start || null,
-        this.shuttle.end || null,
-      ].map((l) => (l && l.toJSDate ? l.toJSDate() : null));
-    },
   },
   watch: {
     currentShuttle() {
@@ -201,18 +187,6 @@ export default {
     },
   },
   methods: {
-    updateDates([start, end], {
-      id, name, firstname, lastname,
-    } = {}) {
-      this.shuttle.driver = {
-        id, name, firstname, lastname,
-      };
-      this.shuttle.start = start instanceof DateTime ? start : DateTime.fromJSDate(start);
-      this.shuttle.end = end instanceof DateTime ? end : DateTime.fromJSDate(end);
-    },
-    updatePassengers(passengers) {
-      this.shuttle.passengers = passengers;
-    },
     async edit(data) {
       const formattedData = {
         ...data,
